Modified Selection Command
Streamlines the select CLI command arguments to make kind
into flags `--functions` and `--layers` flags while also
allowing multiple lookups in a single select command that
are or-ed together. This makes the selection command easier
to use and more flexible, i.e. reduces the need for
as many selection changes.

Also added documentation for the `reviser-shell` entrypoint
that can be used for localized shell execution in situations
like containerized CI.